+++ HarmonyDB.Index/HarmonyDB.Index.Analysis.Tests/Em/TestLoop.cs
@@ -5,7 +5,7 @@ namespace HarmonyDB.Index.Analysis.Tests.Em;
 public class TestLoop : ILoop
 {
     public required string Id { get; init; }
-    public double[,] TonalityProbabilities { get; set; } = new double[Constants.TonicCount, Constants.ScaleCount];
-    public (double TonicScore, double ScaleScore) Score { get; set; }
+    public float[,] TonalityProbabilities { get; set; } = new float[Constants.TonicCount, Constants.ScaleCount];
+    public (float TonicScore, float ScaleScore) Score { get; set; }
     public required (int Tonic, Scale Scale)[] SecretTonalities { get; init; }
 }
