+++ HarmonyDB.Index/HarmonyDB.Index.Analysis/Models/Em/LoopLink.cs
@@ -8,8 +8,8 @@ public record LoopLink : ILoopLink
     public required Song Song { get; init; }
     public string SongId => Song.Id;
     public string LoopId => Loop.Id;
-    public required int Shift { get; init; }
-    public int Weight => (Occurrences + Successions * 4) * (Loop.Length == 2 ? 1 : 5) * (Song.IsTonalityKnown ? 5 : 1);
+    public required byte Shift { get; init; }
+    public float Weight => (Occurrences + Successions * 4) * (Loop.Length == 2 ? 1 : 5) * (Song.IsTonalityKnown ? 5 : 1);
     ISong ILoopLink.Song => Song;
     ILoop ILoopLink.Loop => Loop;
